integrations/TelegramSenderIntegration/requirements.txt
kafka-python
python-telegram-bot
python-dotenv